net: dsa: mv88e6xxx: Add sanity check of received RMU frame
The received RMU frame is expected to be addressed to the MAC address
of the conduit interface. The sequence number should also match. Two
formats codes are accepted. Add these basic sanity checks, and drop
the frame is they are incorrect.  To aid sequence number checking
introduce the use of the DSA inband helpers and its data structures.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>