net: phy: swphy: Support all normal speeds when link down
The software PHY emulator validation function is happy to accept any
link speed if the link is down. swphy_read_reg() however triggers a
WARN_ON(). Change this to report all the standard 1G link speeds are
supported. Once the speed is known the supported link modes will
change, which is a bit odd, but for emulation is probably O.K.

Suggested-by: Russell King (Oracle) <linux@armlinux.org.uk>
Signed-off-by: Andrew Lunn <andrew@lunn.ch>