dsa: mv88e6xxx: Change cmode for disabled ports
Ports 9 and 10 of the 6390 family each have 4 SERDESes. These can
either be used by port itself, or if the port only needs 1 or 2
SERDES, can be used by ports in the range 2-7. This only works if port
9 or 10 release the SERDES by selecting a cmode which does not require
them.

Ensure a disabled port, one which is not listed in DT, has a cmode for
1000baseX, so its SERDES become available for other ports.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>