led: ledtrig-netdev: Get netdev name on register and up.
The PHY containing the LED has a different life cycle to the MAC.  It
is attached to the MAC sometime after probe where the LED class device
is created. It could be the MAC attaches to the PHY during its probe,
i.e. when it registers the netdev, or when open() is called on the
netdev. Make use of the notify to know when these events happen and
ask the PHY for the device it is connected to, so its name can be set.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>