leds: trigger: netdev: introduce validating requested mode
Introduce function to validate the requested mode in preparation for
hw control support. Currently everything is handled in software so
every mode is validated and accepted.

Requested mode are always validated before making any change, to follow
this rework the attr_store function to set the mode to a temp variable
and then apply the new mode only if accepted and validated.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>