dsa: mv88e6xxx: Let 6390X change cmode for lower ports
The 6390X has 8 SERDES ports in total. They are by default assigned to
ports 9 and 10, allowing each to operate at 10GBase-X. However, if
ports 9 and 10 operate at slower speeds requiring less SERDES ports,
they become available for ports 3-8. These lower ports than perform
auto media select, waiting for either the internal PHY or the SERDES
to report link. The driver tries to force the selection to some degree
by disabling/enabling the SERDES. This is performed by faking the
ability to change the CMODE of the port, setting the cache to indicate
the port should be using 1000Base-X for example, and powering up the
SERDES.

This only works if mv88e6390x_port_set_cmode() allows the ports to
change cmode.

Fixes: f39908d3b1c4 ("net: dsa: mv88e6xxx: Set the CMODE for mv88e6390 ports 9 & 10")
Signed-off-by: Andrew Lunn <andrew@lunn.ch>