net: dsa: mv88e6xxx: Set SERDES IRQ as shared
The mv88e6xxx has a single bit to indicate PHY interrupts and SERDES
interrupt. Hence it can be necessary to register two different
interrupt handlers to the same interrupt source. For this to work, the
interrupt needs to be requested as shared. Phylib already does this
for the PHY. Modify the SERDES code to also use shared interrupts.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>