net: bridge: Add static key for net.bridge.bridge-nf-call-iptables
The bridge can invoke netfilter for frames that pass through it,
allowing iptables rules to be applied. By default,
net.bridge.bridge-nf-call-iptables is enabled. However, for embedded
system CONFIG_BRIDGE_NETFILTER is likely to be disabled. Add a static
key to represent net.bridge.bridge-nf-call-iptables, since it needs to
be referenced in the hot path in the following patch.

This change in itself does not fix anything, but provides the
infrastructure for the next patch which is a fix.

Fixes: 472111920f1c ("net: bridge: switchdev: allow the TX data plane forwarding to be offloaded")
Signed-off-by: Andrew Lunn <andrew@lunn.ch>