net: bridge: switchdev: Disable tx fwd offload when netfilter IP enabled
The bridge can offload transmit forwarding to the hardware. The
hardware is passed a single frame, and it will then decided on the
egress interface. For multicast/broadcast frames, the switch will send
the frame out multiple interfaces.

This however breaks in the present of netfilter and
net.bridge.bridge-nf-call-iptables = 1.

One problem is that netfilter only sees a single frame for multicast &
broadcast frames. The egress interface may also be incorrect. As a
result, rules which drop frames based on the egress interface may not
correctly match, or if they do match, drop multicast/broadcast frames
for all egress ports, not a single port.

A second issue is that the result of evaluating if a frames transmit
forward path can be offloaded to the switch is stored in the skbuf
bridge control block data structure. However, when calling into
netfilter iptables, br_validate_ipv4() zeros the IPCB(skb), i.e. IP
control block. However, the bridge and IP control block occupy the
same space in the skb. As a result, the decision to offload the
transmit forward path is lost. The switch will then not offload, but
the bridge only sends a single frame for broadcast/multicast, and as a
result, protocols like DHCP and ARP break.

Extend the rules for when transmit frame forward path offload can be
used such that it is disabled when netfilter iptables are active.

Fixes: 472111920f1c ("net: bridge: switchdev: allow the TX data plane forwarding to be offloaded")
Signed-off-by: Andrew Lunn <andrew@lunn.ch>