drivers: x86: platform: Support for SCU platform
This platform driver is loaded via DMI strings. It instantiates an
bit-banging MDIO bus and a Marvell switch on that bus. A number of I2C
devices are added to an I2C bus. It also exports a number of GPIOs to
userspace and creates a number of GPIO LEDs.