net: dsa: mv88e6xxx: Wait longer after taking out of reset
Before probing for the switch, if there is a reset GPIO, it is
manipulated to take the switch out of reset. A hard coded delay of
1000 to 2000 us is then performed, to allow the switch to get out of
reset.

The time the switch takes to get out of reset actually depends on the
contents of the optional EEPROM, which at reset is read. The function
mv88e6xxx_hardware_reset() polls the switch waiting for it to complete
the reset. However, this cannot be used here, since at this point in
time, it is not known if there is a switch present, or how to access
it.

For the SCU platform, increase the delay to 200ms. Adding logs to
mv88e6xxx_hardware_reset() suggests it takes 120ms.