drivers: leds: ledtrig-netdev: Add firmware binding
Read the configuration of the blink pattern from firmware. This is
done when the LED trigger is activated. The user can change the
configuration afterwards. However, the LED subsystem does not store
any persistent state per trigger. If another trigger is activated, and
then a return to the netdev trigger is made, any user configuration is
lost and the defaults in firmware will be restored.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>