dt-binding: leds: Add netdev trigger properties
The netdev LED trigger can be used to indicate a number of network
conditions, including link, link speed, duplex, transmit activity and
receive activity.  When there is a well defined expected meaning,
e.g. a label on the case next to the LED, DT properties can be used to
configure the trigger to this meaning.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>