net: phy: marvell: Wait longer for cable test to complete
The internal MDIO bus in the Marvell Switch MV88E6390X appears to be
faster than the MDIO bus in MV88E6352. As a result, cable testing with
access to the raw TDR data is timing out because it does not iterate
enough times for the around 14ms a TDR sample takes. Increase the
number of loops to well above what is actually required.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>
Fixes: 0c9bcc1d2394 ("net: phy: marvell: Add support for amplitude graph")