net: r8169: Fixup EEE
phylib now does all the programming of the EEE registers in the
PHY. It also keeps track of all the PHY EEE settings. Remove all PHY
related EEE configuration from this MAC driver.

It is not clear from existing code how to enable/disable the
transmission of LPI when EEE has been negotiated. Make a guess that
the bits being set in rtl*_config_eee_mac() are enabling EEE, and that
clearing the bits will disable the transmission of LPI.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>