net: Add generic support for netdev LEDs
For some devices, the MAC controls the LEDs in the RJ45 connector, not
the PHY. This patchset provides generic support for such LEDs, and
adds the first user, mv88e6xxx.

The common code netdev_leds_setup() is passed a DT node containing the
LEDs and a structure of operations to act on the LEDs. The core will
then create an cdev LED for each LED found in the device tree node.

The callbacks are passed the netdev, and the index of the LED. In
order to make use of this within DSA, helpers are added to convert a
netdev to a ds and port.

The mv88e6xxx has been extended to add basic support for the 6352
LEDs. Only software control is added, but the API supports hardware
offload which can be added to the mv88e6xxx driver later.

For testing and demonstration, the Linksys Mamba aka. wrt1900ac has
the needed DT nodes added to describe its LEDs.

RFC:

netdev is closed at the moment.

A previous version of this patchset was specific to DSA. Vladimir
objected to that and suggested the code should be more generic. Hence
now the shared code is moved into net/core and a struct net_device is
used, rather than struct dsa_switch. As a result, there is a bit more
boilerplate needed in DSA drivers.

If this code is accepted, it would be good to convert the qca8k DSA
driver to also use this framework.

Given how hard it is to get kconfig connect for LEDs, i would not be
surprised if this is also broken :-(

To: Florian Fainelli <f.fainelli@gmail.com>
To: Vladimir Oltean <olteanv@gmail.com>
To: David S. Miller <davem@davemloft.net>
To: Eric Dumazet <edumazet@google.com>
To: Jakub Kicinski <kuba@kernel.org>
To: Paolo Abeni <pabeni@redhat.com>
To: Russell King <linux@armlinux.org.uk>
To: Gregory Clement <gregory.clement@bootlin.com>
Cc: netdev@vger.kernel.org
Signed-off-by: Andrew Lunn <andrew@lunn.ch>

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 1,
    "change-id": "20240316-v6-8-0-net-next-mv88e6xxx-leds-v4-ab77d73d52a4",
    "prefixes": [
      "RFC"
    ]
  }
}