net: mdio: mdio-gpio: Hard code phy_mask
There is a wish to remove the platform data from this driver, since
there are no in-tree users. There are however out of tree users, which
are typically x86 boards, which cannot make use of device tree, and
have a switch attached to the MDIO bus. As such, they don't want the
bus scanned, since there are no PHYs on the bus, just the switch, and
so set the phy_mask to ~0 to stop scanning.

Set this as the hard coded default for the non-DT case. Since there
are no in tree users, there are no backwards compatibility issues.
The next patch will then remove the platform data, breaking the build
for any out of tree drivers, which should act as a heads up, something
has changed.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>