scu: Reimplement using DT overlay
Rather than use platform data to instantiate the Ethernet switch, GPIO
expanders, LEDs, EEPROMs etc, make use of a device tree overlay.

Support for SCU2 and SCU3 has been dropped, in this proof of concept
implementation. This means there is no need to read the nameplate to
determine what devices to instantiate. One overlay creates all the
SCU4 devices. The nameplate is however still exported via
/sysfs. However, the implementation to find the nameplate EEPROM has
changed. When a NVMEM device is registered a notifier chain is
called. Hook into this chain, and look for an NVMEM with name
"nameplate".