gpio: Dynamically add OF nodes if not present
In order that OF overlays can work in a heterogeneous system with
ACPI, ACPI devices need a node in the DT tree so phandles can point to
them, GPIO can be consumed by devices described in DT. Copy the idea
used in CONFIG_PCI_DYNAMIC_OF_NODES to create a dynamic OF node if the
GPIO chip being registered does not already have a node. Connect the
new node to the base of the OF tree.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>