i2c: Add node with phandle and entry to __symbols__ for DT overlays
In order for an I2C bus to be used in a DT overlay, it needs a node in
the device tree. If the bus was not instantiated from device tree, use
OF dynamic to add a node. DT overlays can then add devices to the bus.

Additionally DT overlays resolve phandles via paths, making use of
entries in the root node __symbols__ to translate a node name to a
path. To allow a I2C bus to be resolved when loading an overlay, add
an entry to __symbol__.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>