pci: of: Add phandle and entry to __symbols__ for DT overlays
In order for a PCI device to be used in a DT overlay, it needs an
phandle property. Additionally DT overlays resolve phandles via paths,
making use of entries in the root node __symbols__ to translate a node
name to a path. To allow a PCI device to be resolved when loading an
overlay, add an entry to __symbol__.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>