gpio: gpiolib-of: Add phandle and entry to __symbols__ for DT overlays
In order for a GPIO to be used in a DT overlay, it needs an phandle
property. Additionally DT overlays resolve phandles via paths, making
use of entries in the root node __symbols__ to translate a node name
to a path. To allow a GPIO to be resolved when loading an overlay, add
an entry to __symbol__. The gpiodev device name is not very useful,
"gpiochipX". If the parent is set, use the name from the parent, which
could be something like "gpio_ich", "6-0070", etc.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>