of: resolver: Add helper to add to __symbols__
In order that a device tree overlay can resolve phandles in the base
tree, there needs to be an entry for the node in /__symbols__. Add a
helper dynamic nodes can use to create such entries.

The entries placed in /__symbols__ need to respect the naming
conventions of labels. That means they must start with a letter, and
only contain letters, numbers and _. When creating the label, replace
any unsupported characters with _.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>