net: phy: maxio: Leave core to disable EEE
Calling phy_disable_eee() should be enough to cause phylib to disable
advertisement of EEE. Delete the code writing to the EEE registers.

With that gone, the mmd functions become redundant, so remove
them. They appeared to be the normal C45 over C22, but limited to just
EEE registers. The core version should work just as well.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>