EDITME: cover title for v6.16-rc4-net-next-leds-netns
Make ledtrig-netdev network namespace aware

Interfaces names are associated to a network name space. They are only
unique within a namespace. The current LED trigger netdev does not
take this into account, and so breaks when interfaces are used within
namespaces.

Extend the code so that:

echo "eth42" > /sys/class/leds/phy01/device_name" uses the nets of the
echo process.

When a device registers, compare both the interface name and the
network namespace to make it to a trigger.

When a device changes name, consider the namespace

When a device unregisters from one namespace and registers in another
namespace track this so that it is reassociated to the trigger.

To: Lee Jones <lee@kernel.org>
To: Pavel Machek <pavel@kernel.org>
Cc: linux-leds@vger.kernel.org
Cc: netdev@vger.kernel.org

Signed-off-by: Andrew Lunn <andrew@lunn.ch>

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 1,
    "change-id": "20250707-v6-16-rc4-net-next-leds-netns-b16931c4e7dc",
    "prefixes": [
      "leds"
    ]
  }
}