leds: ledtrig-netdev: Remember netns when cat to /sys/class/leds
The user can associate a netdev to an LED trigger by echo'ing the name
of the netdev to /sys/class/leds/foo/device_name. Record the network
namespace of the process writing to the file, since netdev names are
only unique to a network name space, not globally.

If the trigger already has a reference to a netns, release it. Then
take a reference to the new namespace, if the netdev name is not "",
so preventing the namespace from disappear.

When searching to see if the device exists with the name, search in
the netns.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>