leds: ledtrig-netdev: Handle device changes namespace
A interface can unregister from one namespace and register into a
different name space, potentially with a new name. The netdev pointer
however will remain the same. When an interface associated to a netdev
unregisters, unlink it from the trigger, but keep a reference to its
netdev. If the same netdev registers again, associate it back to the
trigger.

If the user disassociates a interface name from a trigger, also clear
this reference to the netdev, so it does not magically reconnect if it
is moved into a new network namespace.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>