leds: ledtrig-netdev: Handle device changes name
There are two different cases for a device changing name:

The trigger currently is associated to the device. Change the name
stored in the trigger to the devices new name, so keeping the trigger
in sync.

The trigger is not associated to a device, but the new device name,
and its network namespace matches that stored in the trigger. This is
the same as a new device pops into existence, so fall through into
that code.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>