leds: ledtrig-netdev: Handle offloaded device in netns
It is possible to ask an LED which supports offloading of blinking for
a netdev to ask it what netdev it can offload for. Take into account
both the device name and the network namespace the device is in. This
requires modifying the helper set_device_name() to take a namespace as
a parameter. This is then either direct from the interface, or from
the process writing to /sys/class/leds/....

Signed-off-by: Andrew Lunn <andrew@lunn.ch>